programs moved to subfolders did light search. all files re factord correctly.
